SW-4989 Reorg accelerator router code
- moved code into individual subdirs
- extracted out routing for cohorts and deliverables
- replaced AcceleratorMain usage with the 'Page' component which already has built-in header wrappers and snackbar message
- kept the original AcceleratorMain around if we want to use it later
- if there are any styling/spacing issues we can cleanup later
