+++ src/scenes/AcceleratorRouter/Deliverables/DeliverablesList.tsx
@@ -1,19 +1,17 @@
-import React, { useMemo, useRef, useState } from 'react';
+import React, { useMemo, useState } from 'react';
 
 import { Grid, Typography } from '@mui/material';
 import { Separator, TableColumnType } from '@terraware/web-components';
 
 import DeliverablesTable from 'src/components/DeliverablesTable';
+import Page from 'src/components/Page';
 import PageHeader from 'src/components/PageHeader';
 import ParticipantsDropdown from 'src/components/ParticipantsDropdown';
-import PageHeaderWrapper from 'src/components/common/PageHeaderWrapper';
 import { useParticipants } from 'src/hooks/useParticipants';
 import { useLocalization } from 'src/providers';
 import strings from 'src/strings';
 import theme from 'src/theme';
 import { SearchNodePayload } from 'src/types/Search';
-
-import AcceleratorMain from './AcceleratorMain';
 
 const columns = (activeLocale: string | null): TableColumnType[] =>
   activeLocale
@@ -54,7 +52,6 @@ const columns = (activeLocale: string | null): TableColumnType[] =>
 const DeliverablesList = () => {
   const { activeLocale } = useLocalization();
   const { availableParticipants } = useParticipants();
-  const contentRef = useRef(null);
 
   const [participantFilter, setParticipantFilter] = useState<{ participantId?: number }>({ participantId: undefined });
 
@@ -102,19 +99,10 @@ const DeliverablesList = () => {
   );
 
   return (
-    <AcceleratorMain>
-      <PageHeaderWrapper nextElement={contentRef.current}>
-        <PageHeader title={strings.DELIVERABLES} leftComponent={PageHeaderLeftComponent} />
-      </PageHeaderWrapper>
-
+    <Page title={<PageHeader title={strings.DELIVERABLES} leftComponent={PageHeaderLeftComponent} />}>
       {/* -1 for "non-organization scoped search" IE admin search */}
-      <DeliverablesTable
-        columns={columns}
-        extraTableFilters={extraTableFilters}
-        organizationId={-1}
-        pageHeaderRef={contentRef}
-      />
-    </AcceleratorMain>
+      <DeliverablesTable columns={columns} extraTableFilters={extraTableFilters} organizationId={-1} />
+    </Page>
   );
 };
