--- src/scenes/AcceleratorRouter/ModuleContentView.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-import AcceleratorMain from 'src/scenes/AcceleratorRouter/AcceleratorMain';
-
-const ModuleContentView = () => {
-  return <AcceleratorMain />;
-};
-
-export default ModuleContentView;
