+++ src/scenes/AcceleratorRouter/Modules/index.tsx
@@ -1,0 +1,10 @@
+import React from 'react';
+
+import Page from 'src/components/Page';
+import strings from 'src/strings';
+
+const ModuleContentView = () => {
+  return <Page title={strings.MODULE_CONTENT} />;
+};
+
+export default ModuleContentView;
